Problems compiling typescripts in ex
There was some problem in compiling typescript files in ex
folder. It might be because some problem happened in ethers
version. I am so baffled to the version problem in packages
for building and testing the smart contracts. So I deleted
the original project and created new one. The original
version was archived into `exz` folder.